refactor: rename Avatar to AvatarHUD and remove Canvas from DashboardLayout
