+++ src/App.tsx
@@ -3,7 +3,7 @@ import { DashboardUser } from "@/pages/dashboard/DashboardUser";
 import { Skills } from "@/pages/dashboard/Skills";
 import { Profil } from "@/pages/dashboard/Profil";
 import { DashboardLayout } from "@/pages/dashboard/DashboardLayout";
-import { CanvasPage } from "./pages/dashboard/CanvasPage";
+import { CanvasPage } from "./pages/CanvasPage";
 import { Store } from "./pages/dashboard/Store";
 
 function App() {
@@ -14,9 +14,9 @@ function App() {
           <Route index element={<DashboardUser />} />
           <Route path="skills" element={<Skills />} />
           <Route path="profil" element={<Profil />} />
-          <Route path="canvas" element={<CanvasPage />} />
           <Route path="store" element={<Store />} />
         </Route>
+        <Route path="canvas" element={<CanvasPage />} />
       </Routes>
     </BrowserRouter>
   );
